Move messages into a single file
