+++ common/utils/__tests__/api-utils.test.js
@@ -1,14 +1,14 @@
 import { getErrorMessage } from 'common/utils/api-utils';
-import { serverDownErrorMessage } from 'common/constants/api';
+import { networkErrorMessages } from 'common/constants/messages';
 
 describe('API Utilities', () => {
   describe('getErrorMessage', () => {
     it('returns server down if not passed an object', () => {
-      expect(getErrorMessage()).toStrictEqual(serverDownErrorMessage);
+      expect(getErrorMessage()).toStrictEqual(networkErrorMessages.serverDown);
     });
 
     it('returns server down if not passed an object with an expected shape', () => {
-      expect(getErrorMessage({})).toStrictEqual(serverDownErrorMessage);
+      expect(getErrorMessage({})).toStrictEqual(networkErrorMessages.serverDown);
     });
 
     it('returns the passed error message from an expected server error object', () => {
