+++ components/LoginForm/LoginForm.js
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types';
 import { Formik, Field } from 'formik';
 import * as Yup from 'yup';
 import { getErrorMessage } from 'common/utils/api-utils';
-import { minPasswordCharNum, validationErrorMessages } from 'common/constants/validations';
+import { validationErrorMessages } from 'common/constants/messages';
+import { minPasswordCharNum } from 'common/constants/validations';
 import { isMinPasswordStrength } from 'common/utils/validator-utils';
 import Button from 'components/Button/Button';
 import Form from 'components/Form/Form';
